+++ examples/base_1cc_zpa/outputs.tf
@@ -12,45 +12,69 @@ By default, these templates store two critical files to the "examples" directory
    You (and subsequently Zscaler) will NOT be able to remotely access these VMs once deployed without valid SSH access.
 ***Disclaimer***
 
-1) Copy the SSH key to the bastion host
-scp -i ${var.name_prefix}-key-${random_string.suffix.result}.pem ${var.name_prefix}-key-${random_string.suffix.result}.pem ec2-user@${module.bastion.public_dns}:/home/ec2-user/.
+Login Instructions & Resource Attributes
 
-2) SSH to the bastion host
-ssh -i ${var.name_prefix}-key-${random_string.suffix.result}.pem ec2-user@${module.bastion.public_dns}
+CLOUD CONNECTOR Details/Commands:
+SSH to CLOUD CONNECTOR
+%{for k, v in local.cc_map~}
+ssh -F ssh_config ccvm-${k}
+%{endfor~}  
 
-3) SSH to the Cloud Connectors
-ssh -i ${var.name_prefix}-key-${random_string.suffix.result}.pem zsroot@${module.cc_vm.management_ip[0]} -o "proxycommand ssh -W %h:%p -i ${var.name_prefix}-key-${random_string.suffix.result}.pem ec2-user@${module.bastion.public_dns}"
+CLOUD CONNECTOR Management IPs:
+%{for k, v in local.cc_map~}
+ccvm-${k} = ${v}
+%{endfor~}
 
-All CC Management IPs. Replace private IP below with zsroot@"ip address" in ssh example command above.
-${join("\n", module.cc_vm.management_ip)}
+CLOUD CONNECTOR Instance IDs:
+${join("\n", module.cc_vm.id)}
 
-4) SSH to the workload host
-ssh -i ${var.name_prefix}-key-${random_string.suffix.result}.pem ec2-user@${module.workload.private_ip[0]} -o "proxycommand ssh -W %h:%p -i ${var.name_prefix}-key-${random_string.suffix.result}.pem ec2-user@${module.bastion.public_dns}"
+CLOUD CONNECTOR Forwarding/Service IPs:
+${join("\n", module.cc_vm.forwarding_ip)}
 
-All Workload IPs. Replace private IP below with ec2-user@"ip address" in ssh example command above.
-${join("\n", module.workload.private_ip)}
+CLOUD CONNECTOR Forwarding/Service ENIs:
+${join("\n", module.cc_vm.forwarding_eni)}
+
+CLOUD CONNECTOR AZs:
+${join("\n", distinct(module.cc_vm.availability_zone))}
+
+CLOUD CONNECTOR IAM Role ARNs:
+${join("\n", module.cc_iam.iam_instance_profile_arn)}
+
+
+WORKLOAD Details/Commands:
+SSH to WORKLOADS
+%{for k, v in local.workload_map~}
+ssh -F ssh_config workload-${k}
+%{endfor~}  
+
+WORKLOAD IPs:
+%{for k, v in local.workload_map~}
+workload-${k} = ${v}
+%{endfor~} 
+
+WORKLOAD Instance IDs:
+${join("\n", module.workload.instance_id)}
+
+
+BASTION Jump Host Details/Commands:
+1) Copy the SSH key to BASTION home directory
+scp -F ssh_config ${var.name_prefix}-key-${random_string.suffix.result}.pem bastion:~/.
+
+2) SSH to BASTION
+ssh -F ssh_config bastion
+
+BASTION Instance ID:
+${module.bastion.instance_id}
+
 
 VPC:         
 ${module.network.vpc_id}
 
-All CC AZs:
-${join("\n", distinct(module.cc_vm.availability_zone))}
-
-All CC Instance IDs:
-${join("\n", module.cc_vm.id)}
-
-All CC Forwarding/Service IPs:
-${join("\n", module.cc_vm.forwarding_ip)}
-
-All CC Forwarding/Service ENIs:
-${join("\n", module.cc_vm.forwarding_eni)}
+Zscaler Subnet IDs:
+${join("\n", module.network.cc_subnet_ids)}
 
 All NAT GW IPs:
 ${join("\n", module.network.nat_gateway_ips)}
-
-All CC IAM Role ARNs:
-${join("\n", module.cc_iam.iam_instance_profile_arn)}
-
 
 TB
 }
@@ -63,4 +87,45 @@ output "testbedconfig" {
 resource "local_file" "testbed" {
   content  = local.testbedconfig
   filename = "../testbed.txt"
+}
+
+resource "local_file" "ssh_config" {
+  content  = local.ssh_config_contents
+  filename = "../ssh_config"
+}
+
+locals {
+  workload_map = {
+    for index, ip in module.workload.private_ip :
+    index => ip
+  }
+  cc_map = {
+    for index, ip in module.cc_vm.management_ip :
+    index => ip
+  }
+  ssh_config_contents = <<SSH_CONFIG
+    Host bastion
+      HostName ${module.bastion.public_dns}
+      User ec2-user
+      IdentityFile ${var.name_prefix}-key-${random_string.suffix.result}.pem
+    %{for k, v in local.workload_map~}
+Host workload-${k}
+      HostName ${v}
+      User ec2-user
+      IdentityFile ${var.name_prefix}-key-${random_string.suffix.result}.pem
+      StrictHostKeyChecking no
+      ProxyJump bastion
+      ProxyCommand ssh bastion -W %h:%p
+    %{endfor~}
+
+    %{for k, v in local.cc_map~}
+Host ccvm-${k}
+      HostName ${v}
+      User zsroot
+      IdentityFile ${var.name_prefix}-key-${random_string.suffix.result}.pem
+      StrictHostKeyChecking no
+      ProxyJump bastion        
+      ProxyCommand ssh bastion -W %h:%p
+    %{endfor~}
+  SSH_CONFIG
 }