+++ examples/base_ztgateway/outputs.tf
@@ -14,22 +14,6 @@ By default, these templates store two critical files to the "examples" directory
 
 Login Instructions & Resource Attributes
 
-CLOUD CONNECTOR Details/Commands:
-SSH to CLOUD CONNECTOR
-ssh -i ${var.name_prefix}-key-${random_string.suffix.result}.pem zsroot@<< CC mgmt IP >> -o "proxycommand ssh -W %h:%p -i ${var.name_prefix}-key-${random_string.suffix.result}.pem ec2-user@${module.bastion.public_dns}"
-Note: Due to the dynamic nature of autoscaling groups, you will need to login to the AWS console and identify the mgmt IP (network interface device index #1) for each CC deployed and insert into the above command replacing "<< CC mgmt IP >>"
-Note: You can also login to the Cloud Connectors directly from the AWS Console via Session Manager Connect.
-
-All Autoscaling Group IDs:
-${join("\n", module.cc_asg.autoscaling_group_ids)}
-
-Launch Template ID:
-${module.cc_asg.launch_template_id}
-
-CLOUD CONNECTOR IAM Role ARNs:
-${join("\n", module.cc_iam.iam_instance_profile_arn)}
-
-
 WORKLOAD Details/Commands:
 SSH to WORKLOADS
 %{for k, v in local.workload_map~}
@@ -43,6 +27,18 @@ workload-${k} = ${v}
 
 WORKLOAD Instance IDs:
 ${join("\n", module.workload.instance_id)}
+
+WORKLOAD Subnet IDs:
+${join("\n", module.network.workload_subnet_ids)}
+
+WORKLOAD Subnet CIDR Ranges:
+${join("\n", module.network.workload_subnet_cidrs)}
+
+WORKLOAD Subnet Availability Zone Names:
+${join("\n", module.network.workload_subnet_az_names)}
+
+WORKLOAD Subnet Availability Zone ID Mapping:
+${join("\n", module.network.workload_subnet_az_ids)}
 
 
 BASTION Jump Host Details/Commands:
@@ -62,8 +58,14 @@ ${module.network.vpc_id}
 Zscaler Subnet IDs:
 ${join("\n", module.network.cc_subnet_ids)}
 
-All NAT GW IPs:
-${join("\n", module.network.nat_gateway_ips)}
+Zscaler Subnet CIDR Ranges:
+${join("\n", module.network.zs_subnet_cidrs)}
+
+Zscaler Subnet Availability Zone Names:
+${join("\n", module.network.zs_subnet_az_names)}
+
+Zscaler Subnet Availability Zone ID Mapping:
+${join("\n", module.network.zs_subnet_az_ids)}
 
 GWLB Endpoint Service Name:
 ${module.gwlb_endpoint.vpce_service_name}
@@ -73,9 +75,6 @@ ${module.gwlb_endpoint.vpce_service_arn}
 
 All GWLB Endpoint IDs:
 ${join("\n", module.gwlb_endpoint.gwlbe)}
-
-GWLB ARN:
-${module.gwlb.gwlb_arn}
 
 TB
 }