+++ modules/terraform-zscc-asg-aws/versions.tf
@@ -6,11 +6,11 @@ terraform {
     }
     local = {
       source  = "hashicorp/local"
-      version = "~> 2.2.0"
+      version = ">= 2.2.0, < 2.6"
     }
     null = {
       source  = "hashicorp/null"
-      version = "~> 3.1.0"
+      version = "~> 3.1, < 3.3"
     }
   }
   required_version = ">= 0.13.7, < 2.0.0"