+++ modules/terraform-zscc-bastion-aws/outputs.tf
@@ -7,3 +7,7 @@ output "public_dns" {
   description = "Instance Public DNS"
   value       = aws_instance.bastion.public_dns
 }
+output "instance_id" {
+  description = "ID of the EC2 instance"
+  value       = aws_instance.bastion.id
+}